requirements.txt
dash-design-kit==1.6.2
dash-snapshots==1.4.4
dash-user-analytics==0.0.2
gunicorn==20.0.4
openpyxl==3.0.7
psycopg2-binary==2.8.6
pyopenssl==20.0.1
python-language-server==0.35.1
xlrd==2.0.1
dash_bootstrap_components == 0.13.0
scipy == 1.5.4
colorlover == 0.3.0
statsmodels == 0.12.2